Separate custom geometry validation
